python_backend/pyproject.toml
[project]
name = "ai-desk-avatar"
version = "0.1.0"
description = ""
authors = [
    {name = "JiaLong0209",email = "smt02091395@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.12"

dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "ollama (>=0.5.3,<0.6.0)",
    "whisper (>=1.1.10,<2.0.0)",
    "openai-whisper (>=20250625,<20250626)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "gtts (>=2.5.4,<3.0.0)",
    "torch (>=2.8.0,<3.0.0)",
    "numpy (>=1.25,<2.0)",
    "openai-clip (>=1.0.1,<2.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "torchvision (>=0.23.0,<0.24.0)",
    "setuptools (==65.5.0)",
    "ipykernel (>=7.1.0,<8.0.0)",
    "matplotlib (>=3.10.8,<4.0.0)",
    "smplx (>=0.1.28,<0.2.0)",
    "torch-dct (>=0.1.6,<0.2.0)",
    "joblib (>=1.5.3,<2.0.0)",
    "moviepy (==1.0.3)",
    "spacy (>=3.8.11,<4.0.0)",
    "soundfile (>=0.12.1,<1.0.0)",
    "scipy (>=1.16.3,<2.0.0); python_version >= '3.11'",
    "pydub (>=0.25.1,<1.0.0)",
    "pyyaml (>=6.0.1,<7.0.0)",
    "trimesh (>=4.10.1,<5.0.0)",
    "h5py (>=3.15.1,<4.0.0)",
    "pyrender (>=0.1.45,<0.2.0)",
    "unidecode (>=1.4.0,<2.0.0)",
    "pyopenjtalk (>=0.4.1,<0.5.0)",
    "jamo (>=0.4.1,<0.5.0)",
    "pyopenjtalk-prebuilt (>=0.3.0,<0.4.0)",
    "pypinyin (>=0.55.0,<0.56.0)",
    "jieba (>=0.42.1,<0.43.0)",
    "protobuf (>=3.20.2,<6.0.0)",
    "cn2an (>=0.5.23,<0.6.0)",
    "inflect (>=7.5.0,<8.0.0)",
    "eng-to-ipa (>=0.0.2,<0.0.3)",
    "ko-pron (>=1.3,<2.0)",
    "indic-transliteration (==2.3.37)",
    "num-thai (==0.0.5)",
    "opencc (==1.1.1)",
    "demucs (>=4.0.1,<5.0.0)",
    "deep-translator (>=1.11.4,<2.0.0)",
    "hydra-core (>=1.3.2,<2.0.0)",
    "hydra-colorlog (>=1.2.0,<2.0.0)",
    "hydra-optuna-sweeper (>=1.2.0,<2.0.0)",
    "wandb (>=0.24.0,<0.25.0)",
    "rich (>=14.2.0,<15.0.0)",
    "rootutils (>=1.0.7,<2.0.0)",
    "thop (>=0.1.1.post2209072238,<0.2.0)",
    "einops (>=0.8.1,<0.9.0)",
    "lightning (>=2.6.0,<3.0.0)",
    "transformers (==4.42.0)",
    "diffusers (==0.29.0)",
    "google-generativeai (>=0.8.6,<0.9.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "unsloth @ git+https://github.com/unslothai/unsloth.git"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
cython = "^3.2.3"

